refactor: update component imports and restructure package configurations
- Changed imports in layout, page, settings, and template files to use specific component paths.
- Updated TypeScript configuration in desktop and database packages for module resolution.
- Enhanced Storybook configuration to allow access to monorepo packages.
- Refactored package.json files across multiple packages to improve module exports and types.
- Adjusted scripts and dependencies in hooks, storage, types, ui, and validators packages.
- Cleaned up validator exports to remove file extensions.
- Updated turbo.json to include build outputs for various packages.
