+++ apps/desktop/tsconfig.json
@@ -4,9 +4,9 @@
     "tsBuildInfoFile": ".tscache/cache-electron",
     "esModuleInterop": true,
     "jsx": "react",
-    "moduleResolution": "node",
+    "moduleResolution": "nodenext",
     "target": "es2022",
-    "module": "commonjs",
+    "module": "nodenext",
     "outDir": "dist",
     "rootDir": "src",
     "resolveJsonModule": true,
