+++ apps/storybook/.storybook/main.ts
@@ -1,4 +1,8 @@
 import type { StorybookConfig } from '@storybook/react-vite';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.stories.@(ts|tsx)'],
@@ -17,6 +21,18 @@ const config: StorybookConfig = {
   typescript: {
     check: false,
     reactDocgen: false,
+  },
+  viteFinal: async config => {
+    // Allow Vite to access monorepo packages directory for transpilation
+    if (config.server) {
+      config.server.fs = {
+        ...config.server.fs,
+        allow: [
+          path.resolve(__dirname, '../../..'), // Access to monorepo root
+        ],
+      };
+    }
+    return config;
   },
 };
