+++ apps/storybook/src/atoms/Badge.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { Badge } from '@packages/ui';
+import { Badge } from '@packages/ui/atoms/Badge';
 
 /**
  * Badge component for labels and tags with various styles.
