+++ apps/storybook/src/atoms/ProgressBar.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { ProgressBar } from '@packages/ui';
+import { ProgressBar } from '@packages/ui/atoms/ProgressBar';
 
 /**
  * ProgressBar component for displaying progress.
