+++ apps/storybook/src/atoms/RadioGroup.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { RadioGroup } from '@packages/ui';
+import { RadioGroup } from '@packages/ui/atoms/RadioGroup';
 
 /**
  * RadioGroup component for single selection from multiple options.
