+++ apps/storybook/src/atoms/Textarea.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { Textarea } from '@packages/ui';
+import { Textarea } from '@packages/ui/atoms/Textarea';
 
 /**
  * Textarea component for multi-line text input.
