+++ apps/storybook/src/molecules/AvatarGroup.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { AvatarGroup } from '@packages/ui';
+import { AvatarGroup } from '@packages/ui/molecules/AvatarGroup';
 
 /**
  * AvatarGroup component displays multiple avatars with overflow handling.
