+++ apps/storybook/src/molecules/Breadcrumb.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { Breadcrumb } from '@packages/ui';
+import { Breadcrumb } from '@packages/ui/molecules/Breadcrumb';
 
 /**
  * Breadcrumb component for navigation hierarchy.
