+++ apps/storybook/src/organisms/List.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { List } from '@packages/ui';
+import { List } from '@packages/ui/organisms/List';
 
 /**
  * List component for ordered and unordered lists.
