+++ apps/storybook/src/organisms/Menu.stories.tsx
@@ -1,7 +1,7 @@
 // import utils
 import type { Meta, StoryObj } from '@storybook/react';
 // import components
-import { Menu } from '@packages/ui';
+import { Menu } from '@packages/ui/organisms/Menu';
 
 /**
  * Menu component for navigation.
