+++ packages/database/package.json
@@ -2,13 +2,30 @@
     "name": "@packages/database",
     "version": "1.0.0",
     "private": true,
-    "type": "module",
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "main": "./src/index.ts",
+    "module": "./src/index.ts",
+    "types": "./src/index.ts",
+    "sideEffects": false,
+    "files": [
+        "src"
+    ],
     "exports": {
         ".": {
-            "types": "./dist/index.d.ts",
-            "default": "./dist/index.js"
+            "types": "./src/index.ts",
+            "import": "./src/index.ts",
+            "require": "./dist/index.js"
+        },
+        "./*": {
+            "types": "./src/*.ts",
+            "import": "./src/*.ts",
+            "require": "./dist/*.js"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "*": [
+                "./src/*"
+            ]
         }
     },
     "scripts": {
