+++ packages/hooks/package.json
@@ -1,25 +1,46 @@
 {
-  "name": "@packages/hooks",
-  "version": "1.0.0",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./src/index.ts"
+    "name": "@packages/hooks",
+    "version": "1.0.0",
+    "private": true,
+    "type": "module",
+    "main": "./src/index.ts",
+    "module": "./src/index.ts",
+    "types": "./src/index.ts",
+    "sideEffects": false,
+    "files": [
+        "src"
+    ],
+    "exports": {
+        ".": {
+            "types": "./src/index.ts",
+            "default": "./src/index.ts"
+        },
+        "./*": {
+            "types": "./src/*.ts",
+            "default": "./src/*.ts"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "*": [
+                "./src/*"
+            ]
+        }
+    },
+    "scripts": {
+        "build": "echo 'Source-only package - no build needed'",
+        "typecheck": "tsc --noEmit",
+        "lint": "eslint .",
+        "format": "prettier --write .",
+        "format:check": "prettier --check .",
+        "clean": "rm -rf dist .turbo"
+    },
+    "peerDependencies": {
+        "react": "^19.1.0"
+    },
+    "devDependencies": {
+        "@config/typescript": "*",
+        "@types/react": "^19.1.3",
+        "typescript": "^5.8.3"
     }
-  },
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint .",
-    "format": "prettier --write .",
-    "format:check": "prettier --check ."
-  },
-  "dependencies": {
-    "react": "^19.1.0"
-  },
-  "devDependencies": {
-    "@config/typescript": "*",
-    "@types/react": "^19.1.3",
-    "typescript": "^5.8.3"
-  }
 }
