+++ packages/storage/package.json
@@ -1,37 +1,53 @@
 {
-  "name": "@packages/storage",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./src/index.ts"
+    "name": "@packages/storage",
+    "version": "1.0.0",
+    "private": true,
+    "type": "module",
+    "main": "./src/index.ts",
+    "module": "./src/index.ts",
+    "types": "./src/index.ts",
+    "sideEffects": false,
+    "files": [
+        "src"
+    ],
+    "exports": {
+        ".": {
+            "types": "./src/index.ts",
+            "default": "./src/index.ts"
+        },
+        "./*": {
+            "types": "./src/*.ts",
+            "default": "./src/*.ts"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "*": [
+                "./src/*"
+            ]
+        }
+    },
+    "scripts": {
+        "build": "echo 'Source-only package - no build needed'",
+        "typecheck": "tsc --noEmit",
+        "lint": "eslint src",
+        "format": "prettier --write src",
+        "format:check": "prettier --check src",
+        "clean": "rm -rf dist .turbo"
+    },
+    "peerDependencies": {
+        "react": "^19.1.0"
+    },
+    "dependencies": {
+        "mobx": "^6.13.6",
+        "mobx-react-lite": "^4.0.7",
+        "@utils/client": "*"
+    },
+    "devDependencies": {
+        "@config/typescript": "*",
+        "@types/react": "^19.2.0",
+        "typescript": "^5.8.3",
+        "@packages/types": "*",
+        "@packages/validators": "*"
     }
-  },
-  "scripts": {
-    "build": "tsc",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint src",
-    "format": "prettier --write src",
-    "format:check": "prettier --check src",
-    "clean": "rm -rf dist"
-  },
-  "peerDependencies": {
-    "react": "^19.1.0"
-  },
-  "dependencies": {
-    "mobx": "^6.13.6",
-    "mobx-react-lite": "^4.0.7",
-    "@utils/client": "*"
-  },
-  "devDependencies": {
-    "@config/typescript": "*",
-    "@types/react": "^19.2.0",
-    "typescript": "^5.8.3",
-    "@packages/types": "*",
-    "@packages/validators": "*"
-  }
 }
