+++ packages/types/package.json
@@ -4,20 +4,36 @@
     "private": true,
     "type": "module",
     "main": "./src/index.ts",
+    "module": "./src/index.ts",
     "types": "./src/index.ts",
+    "sideEffects": false,
+    "files": [
+        "src"
+    ],
     "exports": {
         ".": {
             "types": "./src/index.ts",
             "default": "./src/index.ts"
+        },
+        "./*": {
+            "types": "./src/*.ts",
+            "default": "./src/*.ts"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "*": [
+                "./src/*"
+            ]
         }
     },
     "scripts": {
-        "build": "tsc",
+        "build": "echo 'Source-only package - no build needed'",
         "typecheck": "tsc --noEmit",
         "lint": "eslint src",
         "format": "prettier --write src",
         "format:check": "prettier --check src",
-        "clean": "rm -rf dist"
+        "clean": "rm -rf dist .turbo"
     },
     "devDependencies": {
         "@config/typescript": "*",
