+++ packages/validators/src/index.ts
@@ -1,5 +1,5 @@
-export * from './app-info.validator.js';
-export * from './open-dialog.validator.js';
-export * from './os-theme.validator.js';
-export * from './settings.validator.js';
-export * from './theme.validator.js';
+export * from './app-info.validator';
+export * from './open-dialog.validator';
+export * from './os-theme.validator';
+export * from './settings.validator';
+export * from './theme.validator';
