+++ packages/validators/tsconfig.json
@@ -2,7 +2,9 @@
   "extends": "@config/typescript/base.json",
   "compilerOptions": {
     "rootDir": "./src",
-    "outDir": "./dist"
+    "outDir": "./dist",
+    "module": "node16",
+    "moduleResolution": "node16"
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist"]
