+++ utils/common/package.json
@@ -3,21 +3,37 @@
     "version": "1.0.0",
     "private": true,
     "type": "module",
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "main": "./src/index.ts",
+    "module": "./src/index.ts",
+    "types": "./src/index.ts",
+    "sideEffects": false,
+    "files": [
+        "src"
+    ],
     "exports": {
         ".": {
-            "types": "./dist/index.d.ts",
-            "default": "./dist/index.js"
+            "types": "./src/index.ts",
+            "default": "./src/index.ts"
+        },
+        "./*": {
+            "types": "./src/*.ts",
+            "default": "./src/*.ts"
+        }
+    },
+    "typesVersions": {
+        "*": {
+            "*": [
+                "./src/*"
+            ]
         }
     },
     "scripts": {
-        "build": "tsc",
+        "build": "echo 'Source-only package - no build needed'",
         "typecheck": "tsc --noEmit",
         "lint": "eslint src",
         "format": "prettier --write src",
         "format:check": "prettier --check src",
-        "clean": "rm -rf dist"
+        "clean": "rm -rf dist .turbo"
     },
     "devDependencies": {
         "@config/typescript": "*",
